+++ src/main/java/model/request/HttpRequestHeader.java
@@ -1,5 +1,7 @@
-package model;
+package model.request;
 
+
+import model.HttpHeader;
 
 import java.util.Map;
