+++ src/test/java/handler/IndexHandlerTest.java
@@ -1,6 +1,6 @@
 package handler;
 
-import model.HttpRequestHeader;
+import model.request.HttpRequestHeader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
