+++ src/test/java/webserver/HandlerSelectorTest.java
@@ -3,7 +3,7 @@ package webserver;
 import handler.IndexHandler;
 import handler.PathHandler;
 import handler.UserHandler;
-import model.HttpRequestHeader;
+import model.request.HttpRequestHeader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
